重构部分数据结构，整理shell.cpp代码结构，初步完成重定位变异在加壳端实现

+++ ParaShell/import.h
@@ -35,13 +35,13 @@ struct Shell_MutatedImpTab_DLLNode
 struct MutatedImpTabInfo
 {
 	void *pMutatedImpTab;
-	unsigned long nMutatedImpTab;
+	DWORD nMutatedImpTab;
 
 	/*
 	TODO:
 	1. “Ï≥£ºÏ≤‚
 	*/
-	MutatedImpTabInfo(unsigned long sz) :
+	MutatedImpTabInfo(DWORD sz) :
 		pMutatedImpTab(0), nMutatedImpTab(sz)
 	{
 		pMutatedImpTab = new char[nMutatedImpTab];
